src/rag-backend/requirements.txt
fastapi
uvicorn[standard]
qdrant-client
google-generativeai
python-dotenv
langchain-text-splitters
unstructured[md]